requirements.txt
fastapi
redis[hiredis]
uvicorn[standard]
gunicorn
pycryptodome
pycryptodomex